chore: Restructure and style fixes
