+++ app/client/src/components/editorComponents/CodeEditor/PeekOverlayPopup/constants.ts
@@ -1,16 +1,3 @@
-export const reactJsonProps = {
-  name: null,
-  enableClipboard: false,
-  displayDataTypes: false,
-  displayArrayKey: true,
-  quotesOnKeys: false,
-  style: {
-    fontSize: "10px",
-  },
-  collapsed: 1,
-  indentWidth: 2,
-  collapseStringsAfterLength: 30,
-};
 export const CONTAINER_MAX_HEIGHT_PX = 252;
 
 export const PEEK_OVERLAY_DELAY = 200;
