+++ app/client/src/components/editorComponents/Debugger/StateInspector/styles.ts
@@ -1,6 +1,5 @@
 import styled, { css } from "styled-components";
 import { Flex, Text } from "@appsmith/ads";
-import { ReactJSONViewerOverrider } from "../../CodeEditor/PeekOverlayPopup/styles";
 
 const imgSizer = css`
   img {
@@ -21,8 +20,4 @@ export const GroupName = styled(Text)`
 
 export const SelectedItem = styled(Flex)`
   ${imgSizer}
-`;
-
-export const JSONWrapper = styled(Flex)<{ $fontSize: string }>`
-  ${ReactJSONViewerOverrider}
 `;
