+++ app/client/src/components/editorComponents/JSONViewer/index.ts
@@ -1,0 +1,3 @@
+export { JSONViewer } from "./JSONViewer";
+export { Size } from "./types";
+export type { JSONViewerProps } from "./types";
