+++ app/client/src/components/editorComponents/JSONViewer/types.ts
@@ -1,0 +1,9 @@
+export enum Size {
+  SMALL = "small",
+  MEDIUM = "medium",
+}
+
+export interface JSONViewerProps {
+  src: unknown;
+  size: Size;
+}
